+++ client/src/app/app.component.spec.ts
@@ -1,27 +1,45 @@
-import { createComponentFactory, Spectator } from '@ngneat/spectator/jest';
-import { RouterTestingModule } from '@angular/router/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { AppComponent } from './app.component';
-import { MatListModule } from '@angular/material/list';
-import { LoadingIndicatorModule } from './modules/loading-indicator/loading-indicator.module';
-import { MatDialog } from '@angular/material/dialog';
-import { SidebarModule } from './modules/sidebar/sidebar.module';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
-import { API_PROJECT_URL, API_PROJECT_DETAIL_URL } from './tokens';
+import { RouterTestingModule } from '@angular/router/testing';
+import { SidebarModule } from './modules/sidebar/sidebar.module';
+import { MatDialog } from '@angular/material/dialog';
+import { API_GENERATE_PROJECT_KEY_URL, API_PROJECT_DETAIL_URL, API_PROJECT_TEMPLATES_URL, API_PROJECT_URL } from './tokens';
+import { MatButtonModule } from '@angular/material/button';
+import { MatIconModule } from '@angular/material/icon';
+import { CUSTOM_ELEMENTS_SCHEMA } from '@angular/core';
+import { StorageService } from './modules/storage/services/storage.service';
 
 describe('AppComponent', () => {
-  const createComponent = createComponentFactory({
-    component: AppComponent,
-    imports: [HttpClientTestingModule, RouterTestingModule, MatListModule, LoadingIndicatorModule, SidebarModule],
-    providers: [MatDialog, { provide: API_PROJECT_DETAIL_URL, useValue: '/api/mock' }, { provide: API_PROJECT_URL, useValue: '/api/mock' }]
-  });
-  let component: any;
-  let spectator: Spectator<AppComponent>;
-  beforeEach(() => {
-    spectator = createComponent();
-    component = spectator.component;
+  let component: AppComponent;
+  let fixture: ComponentFixture<AppComponent>;
+
+  let mockStorageService;
+
+  beforeEach(async () => {
+    mockStorageService = jasmine.createSpyObj(['getItem', 'setItem']);
+    await TestBed.configureTestingModule({
+      declarations: [AppComponent],
+      imports: [HttpClientTestingModule, RouterTestingModule, SidebarModule, MatButtonModule, MatIconModule],
+      providers: [
+        MatDialog,
+        { provide: API_PROJECT_DETAIL_URL, useValue: '/api/mock' },
+        { provide: API_PROJECT_TEMPLATES_URL, useValue: '/api/mock' },
+        { provide: API_GENERATE_PROJECT_KEY_URL, useValue: '/api/mock' },
+        { provide: API_PROJECT_URL, useValue: '/api/mock' },
+        { provide: StorageService, useValue: mockStorageService }
+      ],
+      schemas: [CUSTOM_ELEMENTS_SCHEMA]
+    }).compileComponents();
   });
 
-  it('should create the app', () => {
+  beforeEach(() => {
+    fixture = TestBed.createComponent(AppComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
     expect(component).toBeTruthy();
   });
 });